使用int范围的Hash

+++ src/DistinctCount.java
@@ -1,3 +1,3 @@
 public interface DistinctCount {
-    long compute(int[] element);
+    int compute(int[] element);
 }
